Move the readCredentialsFile helper to value sources
