+++ extension/oauth2clientauthextension/value_sources.go
@@ -3,6 +3,12 @@
 
 package oauth2clientauthextension // import "github.com/open-telemetry/opentelemetry-collector-contrib/extension/oauth2clientauthextension"
 
+import (
+	"fmt"
+	"os"
+	"strings"
+)
+
 type valueSource interface {
 	getValue() (string, error)
 }
@@ -13,6 +19,19 @@ type rawSource struct {
 
 func (v rawSource) getValue() (string, error) {
 	return v.value, nil
+}
+
+func readCredentialsFile(path string) (string, error) {
+	f, err := os.ReadFile(path)
+	if err != nil {
+		return "", fmt.Errorf("failed to read credentials file %q: %w", path, err)
+	}
+
+	credential := strings.TrimSpace(string(f))
+	if credential == "" {
+		return "", fmt.Errorf("empty credentials file %q", path)
+	}
+	return credential, nil
 }
 
 type fileSource struct {
